fix(SLC-109): overall_score NUMERIC + signal-from-activity linking
- DB: overall_score changed from INT to NUMERIC(3,1) for decimal scores
- Migration script: sql/05_fix_overall_score_type.sql
- Signal: createSignalForActivity() links signals to specific activities
- ActivityItem: signal picker button on call/meeting activities
- Signals created from activities now carry activity_id reference

+++ sql/04_v2_migration.sql
@@ -122,7 +122,7 @@ CREATE TABLE IF NOT EXISTS fit_assessments (
   cooperation_score INT,
   conflict_score INT,
   brand_fit_score INT,
-  overall_score INT,
+  overall_score NUMERIC(3,1),
   traffic_light TEXT,
   verdict TEXT,
   reason TEXT,
